requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
asttokens==2.0.4
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==1.5.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.8
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.5
django-cors-headers==3.7.0
django-restframework==0.0.1
djangorestframework==3.12.2
easyocr==1.2.2
entrypoints==0.3
executing==0.5.4
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.2.2
google-api-core==1.24.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-vision==2.0.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
importlib-resources==5.1.2
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.2
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libcst==0.3.16
littleutils==0.2.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.1.6
numpy==1.19.5
nvidia-smi==0.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.8
pandas==0.25.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==6.2.2
pipenv==2020.11.15
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.3
proto-plus==1.13.0
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
pytesseract==0.3.7
pytest==6.2.1
python-bidi==0.4.2
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2020.5
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sorcery==0.2.1
soupsieve==2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tb-nightly==1.14.0a20190603
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.9.1
tesseract==0.1.3
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019060501
tflearn==0.5.0
tifffile==2020.9.3
toml==0.10.2
torchvision==0.4.1
tornado==6.1
tqdm==4.55.1
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.12.1
zipp==3.4.0
